requirements.txt
telethon
google-generativeai
python-dotenv
Pillow
Flask
gunicorn